Implemented The Number of Rows In Section Method on ItemsViewController
Overrode the tableView function numberOfRowsInSection method and within this function I applied a switch statement in order to determine which section to apply Item instances.  The reason for using a switch statement was because an if/else statement seemed to be limiting or could possibly grow into a massive if/else situation based on future features depending on the attributes of an Item instance.